feat: deribit rest api supported

+++ deribit/rest/types/marketdata/get_contract_size.go
@@ -5,5 +5,5 @@ type GetContractSizeParams struct {
 }
 
 type GetContractSizeResponse struct {
-	ContractSize int `json:"contract_size"`
+	ContractSize float64 `json:"contract_size"`
 }
